chg - Код разделен на отдельные модули
