+++ js/consts.js
@@ -1,0 +1,40 @@
+const Comments = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ' Как можно было поймать такой неудачный момент?!',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+];
+
+const Names = [
+  'Иван',
+  'Борис',
+  'Наташа',
+  'Дмитрий',
+  'Саша',
+  'Соня',
+  'Давид',
+  'Юра',
+  'Мария',
+  'Антон',
+];
+
+const Descriptions = [
+  'Ставя палки в колеса, можно остаться и без палки.',
+  'Семь раз промолчи – один раз крякни.',
+  'Одна голова лучше, а два сапога пара!',
+  'Слово не воробей – догони и добей!',
+  'Ломать не строить, пинать не целовать! ',
+  'Упал с самолета – учись летать!',
+  'Кто рано встает, тот не высыпается.',
+  'Будешь мало знать – никогда не состаришься.',
+  'Терпенье и труд мозги перетрут.',
+  'За двумя зайцами погонишься – от обоих схлопочешь! ',
+];
+
+export {Comments, Descriptions, Names};
+
